requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.8.0
httpx>=0.27.0
litellm>=1.40.0
loguru>=0.7.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.2
aiofiles>=23.0.0
python-multipart>=0.0.6
gunicorn>=21.0.0
